Experiment to make expression parsing saner
Honestly I have no idea how I managed to accomplish valid parsing
first time around.

This method comprises two stages. Firstly it bubbles weaker operators as
far up to the front of the node list as possible so that in the second
stage the list can be parsed as unambigous prefix operators.